requirements.txt
pyzotero>=1.5.0
ttkbootstrap>=1.10.0
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
notebooklm-py>=0.1.0
